move files from server directory to root directory
