+++ middlewares/Autorizacao.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/gin-contrib/sessions"
 	"github.com/gin-gonic/gin"
-	"github.com/mquelucci/projeto-loja-virtual/server/controllers/responses"
+	"github.com/mquelucci/projeto-loja-virtual/controllers/responses"
 )
 
 func Auth() gin.HandlerFunc {
